Need state machine for unit component for when they are moving.
