+++ Rpg.Mobile.App/Game/MainBattle/Components/BattleGridScene.cs
@@ -2,6 +2,7 @@
 using Rpg.Mobile.App.Game.MainBattle.Data;
 using Rpg.Mobile.App.Game.MainBattle.Events;
 using Rpg.Mobile.App.Game.MainBattle.StateMachines;
+using Rpg.Mobile.App.Game.MainBattle.StateMachines.Phases;
 using Rpg.Mobile.App.Game.MainBattle.StateMachines.Phases.Setup;
 using Rpg.Mobile.App.Game.UserInterface;
 using Rpg.Mobile.App.Utils;
