+++ Rpg.Mobile.App/Game/MainBattle/StateMachines/Phases/NewRound/NewRoundPhase.cs
@@ -1,6 +1,6 @@
 ﻿using Rpg.Mobile.App.Utils;
 using Rpg.Mobile.GameSdk.Utilities;
-using static Rpg.Mobile.App.Game.MainBattle.StateMachines.BattlePhaseMachine;
+using static Rpg.Mobile.App.Game.MainBattle.StateMachines.Phases.BattlePhaseMachine;
 
 namespace Rpg.Mobile.App.Game.MainBattle.StateMachines.Phases.NewRound;
